# docs/openapi.yaml
openapi: 3.0.3
info:
  title: qf Control Plane API
  description: |
    REST API for the qf host-based firewall control plane.

    **Authentication**: Most endpoints require either a JWT cookie (`qf_token`, set by `POST /auth/login`)
    or an `Authorization: Bearer <token>` header (JWT or opaque API token).

    **Tenant**: Protected endpoints require the `X-Tenant-ID` header (UUID of the tenant).
  version: "0.5.0"
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - cookieAuth: []
  - bearerAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: qf_token
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    TenantID:
      name: X-Tenant-ID
      in: header
      required: true
      description: Tenant UUID
      schema:
        type: string
        format: uuid
    ResourceID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Host:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        hostname:
          type: string
        labels:
          type: object
          additionalProperties:
            type: string
        status:
          type: string
          enum: [pending, enrolling, active, degraded, offline]
        current_generation:
          type: integer
        last_heartbeat_at:
          type: string
          format: date-time
          nullable: true
        agent_version:
          type: string
          nullable: true
        kernel_version:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateHostRequest:
      type: object
      required: [hostname]
      properties:
        hostname:
          type: string
        labels:
          type: object
          additionalProperties:
            type: string
        status:
          type: string
          default: pending
    PatchHostRequest:
      type: object
      properties:
        labels:
          type: object
          additionalProperties:
            type: string
        status:
          type: string

    # ── Policies ───────────────────────────────────────────────────────────
...
    DefaultPolicy:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        default_ingress_action:
          type: string
          enum: [allow, deny]
        default_egress_action:
          type: string
          enum: [allow, deny]
        updated_at:
          type: string
          format: date-time
    PutDefaultPolicyRequest:
      type: object
      properties:
        default_ingress_action:
          type: string
          enum: [allow, deny]
          default: deny
        default_egress_action:
          type: string
          enum: [allow, deny]
          default: deny

    # ── Events ─────────────────────────────────────────────────────────────
...
paths:
...
  /hosts:
    get:
      tags: [Hosts]
      summary: List all hosts
      parameters:
        - $ref: '#/components/parameters/TenantID'
      responses:
        "200":
          description: List of hosts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Host'
        "401":
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Hosts]
      summary: Create a host
      parameters:
        - $ref: '#/components/parameters/TenantID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHostRequest'
      responses:
        "201":
          description: Host created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Host'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /hosts/{id}:
    get:
      tags: [Hosts]
      summary: Get a host by ID
      parameters:
        - $ref: '#/components/parameters/TenantID'
        - $ref: '#/components/parameters/ResourceID'
      responses:
        "200":
          description: Host
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Host'
        "404":
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Hosts]
      summary: Update host labels or status
      parameters:
        - $ref: '#/components/parameters/TenantID'
        - $ref: '#/components/parameters/ResourceID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchHostRequest'
      responses:
        "200":
          description: Updated host
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Host'
        "404":
          $ref: '#/components/responses/NotFound'

  # ── Events ────────────────────────────────────────────────────────────────
...
  /default-policy:
    get:
      tags: [DefaultPolicy]
      summary: Get tenant default ingress/egress actions
      parameters:
        - $ref: '#/components/parameters/TenantID'
      responses:
        "200":
          description: Default policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultPolicy'
    put:
      tags: [DefaultPolicy]
      summary: Set tenant default ingress/egress actions
      parameters:
        - $ref: '#/components/parameters/TenantID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutDefaultPolicyRequest'
      responses:
        "200":
          description: Updated default policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultPolicy'

  # ── Audit Log ─────────────────────────────────────────────────────────────
...